refactor and fix tsx
